Register New Accounts
Register new account with provided account details. Validation is added to make sure the imputs are valid as well as a check to make sure an account with the same email exists in the database.

The password is hashed with a salt using the hash service.

Note: Out of scope, new repo base was added to encapsulate data access.